Variable (comparison, InfoArray etc) is now set up by a perl script rather than macros
This makes it easier to configure the code for each variable.
It also makes error messages easier to understand as it jumps to the code instead of the macro.
Intellisence (MSVC highlighting) is faster to update text coloring and generally understands it better.
Workaround for the issue that C++ doesn't support macros in macros.