Unified the 3 EnumMap classes
Instead of EnumMap, EnumMapInt and EnumMapDefault, now we have only EnumMap with an added parameter.
This new DEFAULT parameter has a default value (0 for ints, -1 for enum types)
EnumMapInt was only used for CityPlotTypes and that is to allow using ints as index argument.
This behavior is now set with a specialized function, meaning it happens automatically.

Having just one class is easier elsewhere where EnumMap is given as an argument.
This is especially true if used in template functions.