Removed the old EnumMap implementation. All instances are now specialized
This means there is no more EnumMap code, which does A if template is this, B is that etc.
Instead the different cases each has a header file of their own, meaning the code is much cleaner.
replaced some defines with the autogenerated VARINFO as this makes compiler errors way more useful.