Added EnumMapCore to allow more compile time optimization
Now the range of valid array entries is set in a base class used by all EnumMaps.
It adds 4 bytes to each EnumMap instance as it caches the size of the array.
This reduces memory I/O by placing it next to the pointer rather than a variable elsewhere.
When the size of the xml file is known at compile time, those variables are replaced with constants.

The real bonus of this is compile time optimization even if the array is too long to make it static.
It only uses static if it's max 64 bools or 4 bytes in total (4 chars, 2 shorts, 1 int)